Moved `EdgeBehavior` `enum` out of `Warp`
