+++ Pinta.Effects/Effects/TileEffect.cs
@@ -250,30 +250,6 @@ public sealed class TileEffect : BaseEffect
 		[Caption ("Edge Behavior")]
 		public EdgeBehavior EdgeBehavior { get; set; } = EdgeBehavior.Wrap;
 	}
-
-	public enum EdgeBehavior
-	{
-		[Caption ("Clamp")]
-		Clamp,
-
-		[Caption ("Wrap")]
-		Wrap,
-
-		[Caption ("Reflect")]
-		Reflect,
-
-		[Caption ("Primary")]
-		Primary,
-
-		[Caption ("Secondary")]
-		Secondary,
-
-		[Caption ("Transparent")]
-		Transparent,
-
-		[Caption ("Original")]
-		Original,
-	}
 }
 
 public enum TileType
